fixed Navlink test
